create general Product.tsx file with all the products in a list which can be accessed by ProductDetail.tsx and ProductOverview.tsx. That way both pages use one single product list which also makes both pages dynamic.
